refactor: update component imports and enhance button styles for consistency
