+++ src/assets/media/svgs/social/facebook-social.vue
@@ -9,7 +9,7 @@
     <rect y="0.5" width="40" height="40" rx="20" fill="white" />
     <path
       d="M29 20.5C29 15.55 24.95 11.5 20 11.5C15.05 11.5 11 15.55 11 20.5C11 25 14.2625 28.7125 18.5375 29.3875V23.0875H16.2875V20.5H18.5375V18.475C18.5375 16.225 19.8875 14.9875 21.9125 14.9875C22.925 14.9875 23.9375 15.2125 23.9375 15.2125V17.4625H22.8125C21.6875 17.4625 21.35 18.1375 21.35 18.8125V20.5H23.825L23.375 23.0875H21.2375V29.5C25.7375 28.825 29 25 29 20.5Z"
-      fill="#5D3BEA"
+      fill="#3686F7"
     />
   </svg>
 </template>
