+++ src/assets/media/svgs/social/instagram.vue
@@ -9,15 +9,15 @@
     <rect y="0.5" width="40" height="40" rx="20" fill="white" />
     <path
       d="M20 13.1265C22.3855 13.1265 22.7108 13.1265 23.6867 13.1265C24.5542 13.1265 24.988 13.3434 25.3133 13.4518C25.747 13.6687 26.0723 13.7771 26.3976 14.1024C26.7229 14.4277 26.9398 14.753 27.0482 15.1867C27.1566 15.512 27.2651 15.9458 27.3735 16.8133C27.3735 17.7892 27.3735 18.006 27.3735 20.5C27.3735 22.994 27.3735 23.2108 27.3735 24.1867C27.3735 25.0542 27.1566 25.488 27.0482 25.8133C26.8313 26.247 26.7229 26.5723 26.3976 26.8976C26.0723 27.2229 25.747 27.4398 25.3133 27.5482C24.988 27.6566 24.5542 27.7651 23.6867 27.8735C22.7108 27.8735 22.494 27.8735 20 27.8735C17.506 27.8735 17.2892 27.8735 16.3133 27.8735C15.4458 27.8735 15.012 27.6566 14.6867 27.5482C14.253 27.3313 13.9277 27.2229 13.6024 26.8976C13.2771 26.5723 13.0602 26.247 12.9518 25.8133C12.8434 25.488 12.7349 25.0542 12.6265 24.1867C12.6265 23.2108 12.6265 22.994 12.6265 20.5C12.6265 18.006 12.6265 17.7892 12.6265 16.8133C12.6265 15.9458 12.8434 15.512 12.9518 15.1867C13.1687 14.753 13.2771 14.4277 13.6024 14.1024C13.9277 13.7771 14.253 13.5602 14.6867 13.4518C15.012 13.3434 15.4458 13.2349 16.3133 13.1265C17.2892 13.1265 17.6145 13.1265 20 13.1265ZM20 11.5C17.506 11.5 17.2892 11.5 16.3133 11.5C15.3374 11.5 14.6867 11.7169 14.1446 11.9337C13.6024 12.1506 13.0602 12.4759 12.5181 13.0181C11.9759 13.5602 11.759 13.994 11.4337 14.6446C11.2169 15.1867 11.1084 15.8374 11 16.8133C11 17.7892 11 18.1145 11 20.5C11 22.994 11 23.2108 11 24.1867C11 25.1627 11.2169 25.8133 11.4337 26.3554C11.6506 26.8976 11.9759 27.4398 12.5181 27.9819C13.0602 28.5241 13.494 28.741 14.1446 29.0663C14.6867 29.2831 15.3374 29.3916 16.3133 29.5C17.2892 29.5 17.6145 29.5 20 29.5C22.3855 29.5 22.7108 29.5 23.6867 29.5C24.6627 29.5 25.3133 29.2831 25.8554 29.0663C26.3976 28.8494 26.9398 28.5241 27.4819 27.9819C28.0241 27.4398 28.241 27.006 28.5663 26.3554C28.7831 25.8133 28.8916 25.1627 29 24.1867C29 23.2108 29 22.8855 29 20.5C29 18.1145 29 17.7892 29 16.8133C29 15.8374 28.7831 15.1867 28.5663 14.6446C28.3494 14.1024 28.0241 13.5602 27.4819 13.0181C26.9398 12.4759 26.506 12.259 25.8554 11.9337C25.3133 11.7169 24.6627 11.6084 23.6867 11.5C22.7108 11.5 22.494 11.5 20 11.5Z"
-      fill="#5D3BEA"
+      fill="#3686F7"
     />
     <path
       d="M20 15.8373C17.3976 15.8373 15.3373 17.8976 15.3373 20.5C15.3373 23.1024 17.3976 25.1627 20 25.1627C22.6024 25.1627 24.6627 23.1024 24.6627 20.5C24.6627 17.8976 22.6024 15.8373 20 15.8373ZM20 23.5361C18.3735 23.5361 16.9639 22.2349 16.9639 20.5C16.9639 18.8735 18.2651 17.4639 20 17.4639C21.6265 17.4639 23.0361 18.7651 23.0361 20.5C23.0361 22.1265 21.6265 23.5361 20 23.5361Z"
-      fill="#5D3BEA"
+      fill="#3686F7"
     />
     <path
       d="M24.7711 16.8133C25.3699 16.8133 25.8554 16.3278 25.8554 15.7289C25.8554 15.1301 25.3699 14.6446 24.7711 14.6446C24.1722 14.6446 23.6867 15.1301 23.6867 15.7289C23.6867 16.3278 24.1722 16.8133 24.7711 16.8133Z"
-      fill="#5D3BEA"
+      fill="#3686F7"
     />
   </svg>
 </template>
