+++ src/assets/media/svgs/social/ok-social.vue
@@ -11,11 +11,11 @@
       fill-rule="evenodd"
       clip-rule="evenodd"
       d="M15.0773 16.1644C15.0773 13.5884 17.3177 11.5 20.0815 11.5C22.8452 11.5 25.0856 13.5884 25.0856 16.1644C25.0856 18.7406 22.8452 20.8291 20.0815 20.8291C17.3177 20.8291 15.0773 18.7406 15.0773 16.1644ZM18.0129 16.1645C18.0129 17.2294 18.939 18.0928 20.0815 18.0928C21.2239 18.0928 22.1501 17.2294 22.1501 16.1645C22.1501 15.0996 21.2239 14.2363 20.0815 14.2363C18.939 14.2363 18.0129 15.0996 18.0129 16.1645Z"
-      fill="#5D3BEA"
+      fill="#3686F7"
     />
     <path
       d="M21.8481 24.5406C23.8515 24.1602 25.052 23.2758 25.1155 23.2284C25.7018 22.7902 25.7959 21.992 25.3258 21.4455C24.8557 20.899 23.9995 20.8113 23.4131 21.2494C23.4007 21.2588 22.1203 22.1743 20.0135 22.1757C17.9068 22.1743 16.5993 21.2588 16.5869 21.2494C16.0005 20.8113 15.1443 20.899 14.6742 21.4455C14.2041 21.992 14.2982 22.7902 14.8845 23.2284C14.9488 23.2765 16.1989 24.1839 18.2584 24.5553L15.3881 27.3514C14.8665 27.8556 14.8822 28.6585 15.4231 29.1448C15.6872 29.3821 16.0275 29.5 16.3675 29.5C16.724 29.5 17.0801 29.3702 17.3471 29.112L20.0136 26.4644L22.9494 29.1292C23.4813 29.6242 24.3427 29.6234 24.8736 29.1278C25.4046 28.6321 25.4039 27.8291 24.8721 27.3342L21.8481 24.5406Z"
-      fill="#5D3BEA"
+      fill="#3686F7"
     />
   </svg>
 </template>
