+++ src/assets/media/svgs/social/telegram-social.vue
@@ -9,7 +9,7 @@
     <rect y="0.5" width="40" height="40" rx="20" fill="white" />
     <path
       d="M12.0384 19.4575C16.9941 17.3406 20.2987 15.9451 21.9521 15.2708C26.6731 13.3457 27.6541 13.0113 28.2935 13.0001C28.4341 12.9978 28.7485 13.032 28.9522 13.194C29.1242 13.3308 29.1715 13.5156 29.1941 13.6453C29.2168 13.7751 29.245 14.0706 29.2226 14.3015C28.9667 16.9368 27.8598 23.3322 27.2966 26.2839C27.0583 27.5328 26.5891 27.9516 26.1348 27.9926C25.1476 28.0816 24.398 27.3529 23.4418 26.7385C21.9457 25.7769 21.1004 25.1783 19.6481 24.24C17.9697 23.1557 19.0577 22.5597 20.0143 21.5857C20.2646 21.3308 24.6142 17.452 24.6984 17.1001C24.7089 17.0561 24.7187 16.892 24.6193 16.8054C24.5199 16.7188 24.3732 16.7484 24.2673 16.772C24.1172 16.8054 21.7271 18.3542 17.0968 21.4186C16.4184 21.8753 15.8038 22.0979 15.2533 22.0862C14.6463 22.0734 13.4787 21.7497 12.6107 21.4731C11.5461 21.1338 10.7 20.9545 10.7737 20.3783C10.8121 20.0781 11.2336 19.7712 12.0384 19.4575Z"
-      fill="#5D3BEA"
+      fill="#3686F7"
     />
   </svg>
 </template>
