+++ src/assets/media/svgs/social/vk-social.vue
@@ -11,7 +11,7 @@
       fill-rule="evenodd"
       clip-rule="evenodd"
       d="M12.9377 14.5H10.7502C10.125 14.5 10 14.7971 10 15.1255C10 15.7106 10.7412 18.6149 13.4539 22.4549C15.2613 25.0782 17.809 26.5 20.1266 26.5C21.5179 26.5 21.69 26.1837 21.69 25.6411V23.6584C21.69 23.0268 21.822 22.9005 22.2621 22.9005C22.5872 22.9005 23.1423 23.0642 24.4417 24.3284C25.925 25.828 26.1691 26.5 27.0033 26.5H29.1919C29.817 26.5 30.1301 26.1837 29.9491 25.5602C29.752 24.9397 29.0438 24.0384 28.1036 22.9702C27.5935 22.3619 26.8293 21.7061 26.5972 21.3776C26.2721 20.9563 26.3652 20.7693 26.5972 20.3944C26.5972 20.3944 29.2629 16.6009 29.541 15.3125C29.681 14.8446 29.541 14.5 28.8788 14.5H26.6912C26.1351 14.5 25.878 14.7971 25.739 15.1255C25.739 15.1255 24.6267 17.8651 23.0503 19.6446C22.5402 20.16 22.3081 20.3247 22.0301 20.3247C21.891 20.3247 21.69 20.16 21.69 19.6921V15.3125C21.69 14.7506 21.5289 14.5 21.0648 14.5H17.6259C17.2788 14.5 17.0698 14.7607 17.0698 15.0083C17.0698 15.5408 17.858 15.6641 17.939 17.1627V20.4176C17.939 21.1321 17.812 21.2614 17.5329 21.2614C16.7917 21.2614 14.9873 18.5088 13.917 15.36C13.7069 14.7476 13.4969 14.5 12.9377 14.5Z"
-      fill="#5D3BEA"
+      fill="#3686F7"
     />
   </svg>
 </template>
