+++ src/components/apps/index.vue
@@ -17,7 +17,7 @@ import SpeechToText from '@/components/apps/speech-to-text/index.vue'
 import Plagiarism from '@/components/apps/plagiarism/index.vue'
 import Plan from '@/components/apps/plan/index.vue'
 import LearningGames from '@/components/apps/learning-games/index.vue'
-import Dashboard from '@/components/apps/apps.vue'
+import Dashboard from '@/components/apps/lists.vue'
 
 const route = useRoute()
