+++ src/components/landing/app-lists/index.vue
@@ -16,8 +16,8 @@
   <div class="my-10 flex items-center justify-center">
     <NuxtLink to="/dashboard">
       <UButton
-        variant="none"
-        class="flex w-fit items-center justify-center bg-[#5D3BEA] text-white lg:w-[420px]"
+        color="blue"
+        class="flex w-fit items-center justify-center rounded-2xl bg-[#5D3BEA] text-white lg:w-[420px]"
       >
         Start Learning now
       </UButton>
